+++ store/mongo/mongo_test.go
@@ -154,35 +154,22 @@ var _ = Describe("Mongo", func() {
 		})
 
 		Context("NewSession", func() {
-			It("returns no error if successful", func() {
-				var err error
-				mongoSession, err = mongoStore.NewSession(log.NewNull())
-				Expect(err).ToNot(HaveOccurred())
+			It("returns a new session if no logger specified", func() {
+				mongoSession = mongoStore.NewSession(nil)
 				Expect(mongoSession).ToNot(BeNil())
+				Expect(mongoSession.Logger()).ToNot(BeNil())
 			})
 
-			It("returns an error if the logger is missing", func() {
-				var err error
-				mongoSession, err = mongoStore.NewSession(nil)
-				Expect(err).To(MatchError("mongo: logger is missing"))
-				Expect(mongoSession).To(BeNil())
-			})
-
-			It("returns an error if the store is closed", func() {
-				mongoStore.Close()
-				Expect(mongoStore.IsClosed()).To(BeTrue())
-				var err error
-				mongoSession, err = mongoStore.NewSession(log.NewNull())
-				Expect(err).To(MatchError("mongo: store closed"))
-				Expect(mongoSession).To(BeNil())
+			It("returns a new session if logger specified", func() {
+				mongoSession = mongoStore.NewSession(logger)
+				Expect(mongoSession).ToNot(BeNil())
+				Expect(mongoSession.Logger()).To(Equal(logger))
 			})
 		})
 
 		Context("with a new session", func() {
 			BeforeEach(func() {
-				var err error
-				mongoSession, err = mongoStore.NewSession(log.NewNull())
-				Expect(err).ToNot(HaveOccurred())
+				mongoSession = mongoStore.NewSession(log.NewNull())
 				Expect(mongoSession).ToNot(BeNil())
 			})